global exceptions handler implemented

+++ src/main/java/com/personaproject/ecommercewebapp/services/CategoryService.java
@@ -3,6 +3,7 @@ package com.personaproject.ecommercewebapp.services;
 import com.personaproject.ecommercewebapp.common.ResponseServices;
 import com.personaproject.ecommercewebapp.dtos.CategoryDTO;
 import com.personaproject.ecommercewebapp.entity.Category;
+import com.personaproject.ecommercewebapp.execeptions.customExceptions.CategoryNotFoundException;
 import com.personaproject.ecommercewebapp.repository.CategoryRepo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -11,7 +12,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -21,19 +21,14 @@ public class CategoryService {
     private final CategoryRepo categoryRepo;
 
     public ResponseEntity<?> createCategory(CategoryDTO categoryDTO) {
-        try {
-            Category category = new Category();
-            category.setCategoryName(categoryDTO.getCategoryName());
-            category.setCategoryDescription(categoryDTO.getDescription());
-            category.setCategoryImageUrl(categoryDTO.getImageUrl());
+        Category category = new Category();
+        category.setCategoryName(categoryDTO.getCategoryName());
+        category.setCategoryDescription(categoryDTO.getDescription());
+        category.setCategoryImageUrl(categoryDTO.getImageUrl());
 
-            categoryRepo.save(category);
-            return new ResponseEntity<>(responseServices.apiResponse(HttpStatus.CREATED, "Category created successfully"), HttpStatus.CREATED);
-//            return responseServices.apiResponse(HttpStatus.CREATED, true, "Category created successfully");
-        } catch (Exception e) {
-//            return responseServices.apiResponse(HttpStatus.BAD_REQUEST, false, "category not created");
-            return new ResponseEntity<>("Category not created", HttpStatus.BAD_REQUEST);
-        }
+        categoryRepo.save(category);
+        return new ResponseEntity<>(responseServices.apiResponse(HttpStatus.CREATED,
+                "Category created successfully"), HttpStatus.CREATED);
     }
 
     private CategoryDTO getCategoryDTO(Category category) {
@@ -44,43 +39,44 @@ public class CategoryService {
         return categoryDTO;
     }
 
+    public void checkIfCategoryExists(Long categoryId) {
+        if (findByID(categoryId) == null)
+            throw new CategoryNotFoundException("Category with ID " + categoryId + " Cannot be found");
+    }
+
     public List<CategoryDTO> listAllCategories() {
         List<Category> categories = categoryRepo.findAll();
-
         ArrayList<CategoryDTO> allCategories = new ArrayList<>();
-
         for (Category category : categories) allCategories.add(getCategoryDTO(category));
         return allCategories;
     }
 
     public CategoryDTO findByID(Long categoryId) {
-        CategoryDTO category = null;
-        Optional<Category> dbCategory = categoryRepo.findById(categoryId);
-        if(dbCategory.isPresent()) {
-             category = getCategoryDTO(dbCategory.get());
-        }
-        return category;
+        if (categoryRepo.findById(categoryId).isEmpty())
+            throw new CategoryNotFoundException("Category with ID " + categoryId + " Cannot be found");
+        return getCategoryDTO(categoryRepo.findById(categoryId).get());
     }
 
     public ResponseEntity<?> updateCategory(Long categoryId, CategoryDTO categoryDTO) {
-        try {
-            Category categoryItem = categoryRepo.getReferenceById(categoryId);
-            categoryItem.setCategoryDescription(categoryDTO.getDescription());
-            categoryItem.setCategoryName(categoryDTO.getCategoryName());
-            categoryItem.setCategoryImageUrl(categoryDTO.getImageUrl());
-            categoryRepo.save(categoryItem);
-            return new ResponseEntity<>(responseServices.apiResponse(HttpStatus.ACCEPTED,
-                    "Category with ID " + categoryId + " updated!"), HttpStatus.ACCEPTED);
-        } catch (Exception e) {
-            return new ResponseEntity<>(responseServices.apiResponse(HttpStatus.BAD_REQUEST,
-                    "Category cannot be updated"), HttpStatus.BAD_REQUEST);
-        }
+        if (categoryRepo.findById(categoryId).isEmpty())
+            throw new CategoryNotFoundException("Category with ID " + categoryId + " Cannot be found");
+
+        Category categoryItem = categoryRepo.getReferenceById(categoryId);
+        categoryItem.setCategoryDescription(categoryDTO.getDescription());
+        categoryItem.setCategoryName(categoryDTO.getCategoryName());
+        categoryItem.setCategoryImageUrl(categoryDTO.getImageUrl());
+        categoryRepo.save(categoryItem);
+        return new ResponseEntity<>(responseServices.apiResponse(HttpStatus.ACCEPTED,
+                "Category with ID " + categoryId + " updated!"), HttpStatus.ACCEPTED);
     }
 
-    public ResponseEntity<?> removeCategory(Long categoryId) {
+    public ResponseEntity<?> removeCategory(Long categoryId) throws CategoryNotFoundException {
+        if (categoryRepo.findById(categoryId).isEmpty())
+            throw new CategoryNotFoundException("Category with ID " + categoryId + " Cannot be found");
+
         categoryRepo.deleteById(categoryId);
         return new ResponseEntity<>(responseServices.apiResponse(HttpStatus.OK,
-                "Category with ID " + categoryId + " deleted!"), HttpStatus.BAD_REQUEST);
+                "Category with ID " + categoryId + " deleted!"), HttpStatus.OK);
     }
 
     public Object removeAllCategory() {
